yktrace: move all SIR related stuff into a new sir.rs file.
I found it confusing that the SIR was stored in tir.rs and lots of
sir-related stuff was also in lib.rs.
